requirements.txt
pyserial
lxml